feat(settings): extract the Advanced pages, unfile two rows
Why: completes the registry — PAGES loses its cast, so Record
<SettingsPageId, …> now makes a missing page a compile error.

'Watch the working copy' and 'Terminal shell' finally leave Pull & fetch,
which described neither, for a Workspace page that does. Settings.tsx
holds no cards of its own any more.
